requirements.txt
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.2
future==0.18.2
imageio==2.9.0
imgaug==0.4.0
joblib==0.16.0
kiwisolver==1.2.0
matplotlib==3.3.1
networkx==2.5
nose==1.3.7
numpy==1.19.2
opencv-python==4.4.0.42
Pillow==7.2.0
prefetch-generator==1.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
Shapely==1.7.1
six==1.15.0
tensorly==0.4.5
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.6.0
torchvision==0.7.0
tqdm==4.49.0
gdown=4.7.1
pandas=2.0.1